Backend/requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.4
Django==3.2.6
django-ckeditor==6.1.0
django-cookies-samesite==0.9.0
django-cors-headers==3.8.0
django-environ==0.4.5
django-js-asset==1.2.2
djangorestframework==3.12.4
drf-extensions==0.7.1
drf-nested-routers==0.93.3
idna==3.2
mysqlclient==2.0.3
Pygments==2.10.0
pytz==2021.1
requests==2.26.0
sqlparse==0.4.1
ua-parser==0.10.0
urllib3==1.26.6
